Check for pre-existing details before updating in the API.

+++ src/controller/user.js
@@ -1,44 +1,36 @@
-const prisma = require("../config/DbConfig")
-const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
-const SECRET = process.env.SECRET
-const fs = require('fs');
-const path = require('path');
-const { v4: uuidv4 } = require('uuid');
+const prisma = require("../config/DbConfig");
+const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
+const SECRET = process.env.SECRET;
+const fs = require("fs");
+const path = require("path");
+const { v4: uuidv4 } = require("uuid");
 const { ROLES } = require("../utills/constant");
 
 const createUser = async (req, h) => {
     try {
-        const { name, email, contact_no, location, password, role_id } = req.payload;
+        const { name, email, contact_no, password, role_id } = req.payload;
 
         const existingUser = await prisma.user.findFirst({
             where: {
-                email: email,
+                OR: [{ email: email }, { contact_no: contact_no }],
                 deleted_at: null,
             },
         });
 
         if (existingUser) {
-            return h.response({ message: "This user already exists. Please login." }).code(400);
+            return h
+                .response({
+                    message: "This user already exists. Please login.",
+                })
+                .code(400);
         }
-
-        // file upload
-        const { profile_image: file } = req.payload;
-        const uploadDir = path.join(__dirname, "..", "uploads");
-        const uniqueFilename = `${Date.now()}_${file.hapi.filename}`;
-        const uploadPath = path.join(uploadDir, uniqueFilename);
-
-        if (!fs.existsSync(uploadDir)) {
-            fs.mkdirSync(uploadDir, { recursive: true });
-        }
-        const fileStream = fs.createWriteStream(uploadPath);
-        file.pipe(fileStream);
 
         const roleExists = await prisma.role.findFirst({
             where: {
                 id: role_id,
-                deleted_at: null
-            }
+                deleted_at: null,
+            },
         });
         if (!roleExists) {
             return h.response({ message: "Role not found." }).code(404);
@@ -47,25 +39,33 @@ const createUser = async (req, h) => {
         const uuidCode = uuidv4().substring(0, 5);
         let usercode;
 
-        if (roleExists.role === ROLES.RESTAURANT_OWNER) {
-            usercode = `OW${uuidCode}`
-        } else if (roleExists.role === ROLES.USER) {
-            usercode = `US${uuidCode}`
+        if (roleExists.role === ROLES.USER) {
+            usercode = `US${uuidCode}`;
         } else if (roleExists.role === ROLES.SUPER_ADMIN) {
-            usercode = `SA${uuidCode}`
+            usercode = `SA${uuidCode}`;
         } else if (roleExists.role === ROLES.ADMIN) {
-            usercode = `AD${uuidCode}`
+            usercode = `AD${uuidCode}`;
         }
 
         const hashedPassword = await bcrypt.hash(password, 10);
         const user = await prisma.user.create({
             data: {
-                name, email, contact_no, location, password: hashedPassword, role_id: roleExists.id,
-                profile_image: uniqueFilename, usercode,
+                name,
+                email,
+                contact_no,
+                password: hashedPassword,
+                role_id: roleExists.id,
+                usercode,
             },
         });
 
-        return h.response({ success: true, message: "User created successfully.", data: user }).code(201);
+        return h
+            .response({
+                success: true,
+                message: "User created successfully.",
+                data: user,
+            })
+            .code(201);
     } catch (error) {
         console.log(error);
         return h.response({ message: "Something went wrong", error }).code(500);
@@ -95,10 +95,10 @@ const userLogin = async (req, h) => {
                     select: {
                         id: true,
                         role: true,
-                    }
+                    },
                 },
                 created_at: true,
-            }
+            },
         });
         if (!user) {
             return h.response({ message: "User not found" }).code(404);
@@ -109,11 +109,16 @@ const userLogin = async (req, h) => {
             return h.response({ message: "Invalid password" }).code(400);
         }
         const token = jwt.sign({ email: user.email }, SECRET, {
-            expiresIn: "1d"
+            expiresIn: "1d",
         });
 
-        return h.response({ message: "Login sucessfully", token: token, data: user }).code(200);
-
+        return h
+            .response({
+                message: "Logged in sucessfully.",
+                token: token,
+                data: user,
+            })
+            .code(200);
     } catch (error) {
         console.log(error);
         return h.response({ message: "Error while login", error }).code(500);
@@ -145,19 +150,24 @@ const me = async (req, h) => {
                     select: {
                         id: true,
                         role: true,
-                    }
+                    },
                 },
                 created_at: true,
                 updated_at: true,
-            }
+            },
         });
 
-        return h.response({ message: "User profile_image fetched successfully", data: user }).code(200);
+        return h
+            .response({
+                message: "User profile_image fetched successfully",
+                data: user,
+            })
+            .code(200);
     } catch (error) {
         console.log(error);
         return h.response({ message: "Something went wrong", error }).code(500);
     }
-}
+};
 
 // edit user profile
 const editMyProfile = async (req, h) => {
@@ -173,13 +183,37 @@ const editMyProfile = async (req, h) => {
             },
         });
 
+        const checkSame = [];
+        if (name && name === existingUser.name) checkSame.push("Name");
+        if (contact_no && contact_no === existingUser.contact_no)
+            checkSame.push("Contact number");
+        if (location && location === existingUser.location)
+            checkSame.push("Address");
+        if (checkSame.length > 0) {
+            return h
+                .response({
+                    success: false,
+                    message: `${checkSame.join(
+                        ", "
+                    )} match the current details in our database. Please modify these fields or leave them blank if no changes are necessary.`,
+                })
+                .code(400);
+        }
+
         let uniqueFilename;
-        if (profile_image && profile_image.hapi && profile_image.hapi.filename) {
+        if (
+            profile_image &&
+            profile_image.hapi &&
+            profile_image.hapi.filename
+        ) {
             const uploadDir = path.join(__dirname, "..", "uploads");
 
             // Delete existing profile image if it exists
             if (existingUser && existingUser.profile_image) {
-                const existingImagePath = path.join(uploadDir, existingUser.profile_image);
+                const existingImagePath = path.join(
+                    uploadDir,
+                    existingUser.profile_image
+                );
                 if (fs.existsSync(existingImagePath)) {
                     fs.unlinkSync(existingImagePath);
                 }
@@ -204,23 +238,28 @@ const editMyProfile = async (req, h) => {
                 name,
                 contact_no,
                 location,
-                profile_image: uniqueFilename || existingUser.profile_image
-            }
+                profile_image: uniqueFilename || existingUser.profile_image,
+            },
         });
 
-        return h.response({ success: true, message: "Profile updated successfully.", data: updatedData }).code(200);
-
+        return h
+            .response({
+                success: true,
+                message: "Profile updated successfully.",
+                data: updatedData,
+            })
+            .code(200);
     } catch (error) {
         console.log(error);
-        return h.response({ message: "Error while editing user's profile" }).code(500);
+        return h
+            .response({ message: "Error while editing user's profile" })
+            .code(500);
     }
-}
-
+};
 
 module.exports = {
     createUser,
     userLogin,
     me,
     editMyProfile,
-
-}
+};
